Restructure repository and add architecture docs
