+++ Vanta/__init__.py
@@ -1,0 +1,13 @@
+"""Top-level Vanta package."""
+from .core.UnifiedVantaCore import UnifiedVantaCore, get_vanta_core
+from .core.UnifiedAgentRegistry import UnifiedAgentRegistry
+from .core.UnifiedAsyncBus import UnifiedAsyncBus
+from .core.UnifiedMemoryInterface import UnifiedMemoryInterface
+
+__all__ = [
+    "UnifiedVantaCore",
+    "get_vanta_core",
+    "UnifiedAgentRegistry",
+    "UnifiedAsyncBus",
+    "UnifiedMemoryInterface",
+]
