+++ Vanta/core/__init__.py
@@ -1,0 +1,19 @@
+"""Core modules for Vanta."""
+from .UnifiedVantaCore import UnifiedVantaCore, get_vanta_core
+from .UnifiedAgentRegistry import UnifiedAgentRegistry
+from .UnifiedAsyncBus import UnifiedAsyncBus
+from .UnifiedMemoryInterface import UnifiedMemoryInterface
+from .VantaBLTMiddleware import VantaBLTMiddleware
+from .VantaCognitiveEngine import VantaCognitiveEngine
+from .VantaOrchestrationEngine import VantaOrchestrationEngine
+
+__all__ = [
+    "UnifiedVantaCore",
+    "get_vanta_core",
+    "UnifiedAgentRegistry",
+    "UnifiedAsyncBus",
+    "UnifiedMemoryInterface",
+    "VantaBLTMiddleware",
+    "VantaCognitiveEngine",
+    "VantaOrchestrationEngine",
+]
