+++ Vanta/integration/__init__.py
@@ -1,0 +1,1 @@
+"""Integration components."""
