+++ __init__.py
@@ -4,15 +4,15 @@ VoxSigil GUI Components - UI components for the GridFormer and Vanta systems
 
 # Import all interface components for easy access
 # Note: Real DynamicGridFormerGUI is now fixed and can be imported
-from .dynamic_gridformer_gui import DynamicGridFormerGUI
+from .gui.dynamic_gridformer_gui import DynamicGridFormerGUI
 
 # from .model_discovery_interface import ModelDiscoveryInterface
-# from .model_tab_interface import VoxSigilModelInterface
+# from .gui.model_tab_interface import VoxSigilModelInterface
 # from .neural_interface import NeuralInterface
-# from .performance_tab_interface import VoxSigilPerformanceInterface
-# from .testing_tab_interface import VoxSigilTestingInterface
+# from .gui.performance_tab_interface import VoxSigilPerformanceInterface
+# from .gui.testing_tab_interface import VoxSigilTestingInterface
 # from .training_interface import VoxSigilTrainingInterface
-# from .visualization_tab_interface import VoxSigilVisualizationInterface
+# from .gui.visualization_tab_interface import VoxSigilVisualizationInterface
 # from .visualization_utils import GridVisualizer, PerformanceVisualizer
 # Import new enhanced components that are working
 from .enhanced_testing_interface import EnhancedVoxSigilTestingInterface
