+++ docs/architecture/system_modules.puml
@@ -1,0 +1,32 @@
+@startuml
+!theme plain
+title Vanta System Module Overview
+
+package "Vanta Core" {
+  [UnifiedVantaCore]
+  [UnifiedAgentRegistry]
+  [UnifiedAsyncBus]
+  [UnifiedMemoryInterface]
+}
+package Handlers {
+  [RagIntegrationHandler]
+  [SpeechIntegrationHandler]
+  [VMBIntegrationHandler]
+}
+package GUI {
+  [DynamicGridFormerGUI]
+  [ModelTab]
+  [PerformanceTab]
+  [TestingTab]
+  [VisualizationTab]
+}
+[UnifiedVantaCore] --> [UnifiedAgentRegistry]
+[UnifiedVantaCore] --> [UnifiedAsyncBus]
+[UnifiedVantaCore] --> [UnifiedMemoryInterface]
+[UnifiedVantaCore] --> Handlers
+Handlers --> [DynamicGridFormerGUI]
+[DynamicGridFormerGUI] --> [ModelTab]
+[DynamicGridFormerGUI] --> [PerformanceTab]
+[DynamicGridFormerGUI] --> [TestingTab]
+[DynamicGridFormerGUI] --> [VisualizationTab]
+@enduml
