+++ handlers/__init__.py
@@ -1,0 +1,10 @@
+"""Integration handlers package."""
+from .rag_integration_handler import RagIntegrationHandler
+from .speech_integration_handler import SpeechIntegrationHandler
+from .vmb_integration_handler import VMBIntegrationHandler
+
+__all__ = [
+    "RagIntegrationHandler",
+    "SpeechIntegrationHandler",
+    "VMBIntegrationHandler",
+]
