+++ handlers/speech_integration_handler.py
