+++ handlers/vmb_integration_handler.py
