+++ scripts/vmb_final_demo.py
